fix: 修复UPSERT冲突条件与索引定义不匹配问题
- ON CONFLICT条件从COALESCE(upgrade_package_id, 0)改为COALESCE(upgrade_package_id, (0)::bigint)
- 与索引定义uq_du_device_firmware_package保持一致
- 解决升级记录无法正确关联到任务的问题

+++ inv_api_server/internal/repository/ota_repository.go
@@ -1374,7 +1374,7 @@ func (r *OTARepository) CreateTaskFromAppTrigger(ctx context.Context, userID int
 			    old_version, status, progress, error_message, retry_count, pushed_by,
 			    upgrade_package_id, task_id, source, created_at, updated_at)
 			VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, NOW(), NOW())
-			ON CONFLICT (device_sn, firmware_id, COALESCE(upgrade_package_id, 0))
+			ON CONFLICT (device_sn, firmware_id, COALESCE(upgrade_package_id, (0)::bigint))
 			DO UPDATE SET
 			    status = EXCLUDED.status,
 			    task_id = EXCLUDED.task_id,
